requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
python-dotenv
ipykernel
streamlit
openai
diffusers 
transformers 
torch